File structure, documentation and additional fixes
