+++ framework/bucket_policies/access_validation_strategies/access_validation_strategy_interface.py
